Fix AspectRatio only returning integer values

+++ Libraries/SpriteTools/Code/TextureAtlas.cs
@@ -20,7 +20,7 @@ public class TextureAtlas
     /// <summary>
     /// Returns the aspect ratio of a frame from the texture atlas.
     /// </summary>
-    public float AspectRatio => (MaxFrameSize.y == 0) ? ((Texture.Height == 0) ? 1 : (Texture.Width / Texture.Height)) : (MaxFrameSize.x / MaxFrameSize.y);
+    public float AspectRatio => (MaxFrameSize.y == 0) ? ((Texture.Height == 0) ? 1 : ((float)Texture.Width / Texture.Height)) : (MaxFrameSize.x / MaxFrameSize.y);
 
     /// <summary>
     /// Returns the UV tiling for the texture atlas.
